create a plugin to update the spec title element
