+++ profiles/dini.js
@@ -16,6 +16,7 @@ const modules = [
   import("../src/core/data-include.js"),
   import("../src/core/markdown.js"),
   import("../src/core/reindent.js"),
+  import("../src/core/title.js"),
   import("../src/dini/headers.js"),
   import("../src/dini/abstract.js"),
   import("../src/core/data-transform.js"),
