+++ profiles/w3c-common.js
@@ -34,6 +34,7 @@ const modules = [
   import("../src/core/data-include.js"),
   import("../src/core/markdown.js"),
   import("../src/core/reindent.js"),
+  import("../src/core/title.js"),
   import("../src/w3c/headers.js"),
   import("../src/w3c/abstract.js"),
   import("../src/core/data-transform.js"),
